changed the server to generate unique friend codes
The unique friend codes will be used instead of IP addresses for
looking up if someone in the server. The unique ID's will be generated
for a client immediately upon connecting to the server. Then the server
will send that ID to the client so they can give it out as a "friend
code" to other players for them to use to request to play chess with
instead of using their IP address.